refactored Paul's additions for compilation and previous test passing

+++ tests/test-circle.cpp
@@ -14,7 +14,7 @@ using std::pair;
 TEST_CASE("test circle shape") {
   SECTION("constructors behave") {
     unsigned int radius = 5;
-    pair<unsigned int, unsigned int> bound_box = make_pair(10, 10);
+    pair<float, float> bound_box = make_pair(10, 10);
     pair<unsigned int, unsigned int> current_point = make_pair(0, 0);
 
     auto test_circle = Circle(radius, bound_box, current_point);
